moved types to their own
