+++ src/medimail/medimail-client.service.ts
@@ -7,144 +7,10 @@ import {
 } from 'soap';
 import { ISoapError } from 'soap/lib/client';
 import { WSDL } from 'soap';
+import { CheckboxReturn, CheckboxType, FormattedMSSResponse, HelloResult, MSSSoapResult, MedimailActions, OpenResult, SendMessageOptions, SendResult } from 'src/medimail/models/medimail.types';
 
 const formatDateToDatetime = (date: Date): string => date.toISOString().slice(0, 19);
 
-export enum MedimailActions {
-  HELLO = 'hello',
-  SEND = 'send',
-  OPEN = 'open',
-  CHECKBOX = 'checkbox',
-}
-
-type Attachment = {
-  name: string;
-  content: string;
-};
-
-type SendAttachments = {
-  attachment1?: Attachment;
-  attachment2?: Attachment;
-  attachment3?: Attachment;
-  attachment4?: Attachment;
-  attachment5?: Attachment;
-};
-
-type MSSSoapResult = {
-  return: {
-    $value: string;
-  };
-  sendResult: {
-    $value: string;
-  };
-  openReturn: {
-    $value: string;
-  };
-  checkboxReturn: {
-    $value: string;
-  };
-};
-
-type FormattedMSSResponse = {
-  formatted: MSSSoapResult;
-  xml: string;
-  error: Error | null;
-};
-
-type SendMessageOptions = {
-  title: string;
-  message: string;
-  signatories: string[];
-  recipients?: string[];
-  attachments?: SendAttachments;
-};
-
-type HelloResult = {
-  hello: {
-    status: string;
-    hello: string;
-  };
-};
-
-type SendResult = {
-  webisend: {
-    status: string,
-    author: string,
-    signatories: string,
-    ignore: null,
-    title: string,
-    refs: {
-      mss: string[] | string
-    },
-  }
-}
-
-type Signatory = {
-  firstname: string;
-  lastname: string;
-  email: string;
-}
-
-type Recipient = {
-  firstname: string;
-  lastname: string;
-  email: string;
-}
-
-type OpenResult = {
-  webiopen: {
-    status: string,
-    call: string,
-    author: {
-      firstname: string,
-      lastname: string,
-      email: string,
-    },
-    begindate: string,
-    enddate: string | null,
-    signatories: Signatory[],
-    recipients: Recipient[],
-    title: string,
-    content: string,
-    nbattachments: string,
-    attachments: Attachment[],
-    'message-id': string,
-    inReplyTo: string | null,
-    references: string | null,
-    replyTo: string,
-  }
-}
-
-type CheckboxMessage = {
-  ref: string,
-  author: Recipient,
-  dtcreate: string,
-  signatories: Signatory[],
-  recipients: null,
-  title: string,
-}
-
-
-type CheckboxReturn = {
-  webicheckbox: {
-    status: string,
-    user: string,
-    begindate: string,
-    enddate: string,
-    inputs: {
-      kvp: CheckboxMessage[]
-    },
-    outputs: {
-      kvp: CheckboxMessage[]
-    },
-  }
-}
-
-export enum CheckboxType {
-  ALL_MESSAGES = 1,
-  RECEIVED_MESSAGES = 2,
-  SENT_MESSAGES = 3,
-}
 
 export class MedimailClient {
   private readonly _wsdlUrl: string = path.resolve(
